Refactored: Renamed 'IpcClient/Server' into 'BackgroundClient/Service'
